new(demo): refactor glyphs to Codesandbox example
